locking down delete item

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -50,6 +50,11 @@ const Mutations = {
     // 1. Find the item
     const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
     // 2. Check if they own that item, or have permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+    if(!ownsItem && !hasPermissions) {
+      throw new Error('You don\'t have the required permissions to do that!');
+    }
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
